Contexts of a context tree

This function extracts from a context tree a description of all of its contexts.

Usage

# S3 method for class 'ctx_tree'
contexts(
  ct,
  sequence = FALSE,
  reverse = FALSE,
  frequency = NULL,
  positions = FALSE,
  ...
)

# S3 method for class 'ctx_tree_cpp'
contexts(
  ct,
  sequence = FALSE,
  reverse = FALSE,
  frequency = NULL,
  positions = FALSE,
  ...
)

Arguments

ct

a context tree.

sequence

if TRUE the function returns its results as a data.frame, if FALSE (default) as a list of ctx_node objects. (see details)

reverse

logical (defaults to FALSE). See details.

frequency

specifies the counts to be included in the result data.frame. The default value of NULL does not include anything. "total" gives the number of occurrences of each context in the original sequence. "detailed" includes in addition the break down of these occurrences into all the possible states.

positions

logical (defaults to FALSE). Specify whether the positions of each context in the time series used to build the context tree should be reported in a positions column of the result data frame. The availability of the positions depends on the way the context tree was built. See details for the definition of a position.

...

additional arguments for the contexts function.

Value

A list of class contexts containing the contexts represented in this tree (as ctx_node) or a data.frame.

Details

The default behaviour of the function is to return a list of all the contexts using ctx_node objects (as returned by find_sequence()). The properties of the contexts can then be explored using adapted functions such as counts() and positions().

When sequence=TRUE the method returns a data.frame whose first column, named context, contains the contexts as vectors (i.e. the value returned by as_sequence() applied to a ctx_node object). Other columns contain context specific values specified by the additional parameters. Setting any of those parameters to a value that ask for reporting information will toggle the result type of the function to data.frame.

If frequency="total", an additional column named freq gives the number of occurrences of each context in the series used to build the tree. If frequency="detailed", one additional column is added per state in the context space. Each column records the number of times a given context is followed by the corresponding value in the original series.

Positions

A position of a context ctx in the time series x is an index value t such that the context ends with x[t]. Thus x[t+1] is after the context. For instance if x=c(0, 0, 1, 1) and ctx=c(0, 1) (in standard state order), then the position of ctx in x is 3.

State order in a context

Notice that contexts are given by default in the temporal order and not in the "reverse" order used by many VLMC research papers: older values are on the left. For instance, the context c(1, 0) is reported if the sequence 0, then 1 appeared in the time series used to build the context tree. Set reverse to TRUE for the reverse convention which is somewhat easier to relate to the way the context trees are represented by draw() (i.e. recent values at the top the tree).

See also

Examples

dts <- sample(as.factor(c("A", "B", "C")), 100, replace = TRUE)
dts_tree <- ctx_tree(dts, max_depth = 3, min_size = 5)
## direct representation with ctx_node objects
contexts(dts_tree)
#> Contexts:
#>  A, A
#>  B, A
#>  B, C, A
#>  C, A
#>  A, B
#>  B, B, B
#>  C, B, B
#>  B, B
#>  B, C, B
#>  C, C, B
#>  C, B
#>  C, A, C
#>  A, C
#>  A, B, C
#>  C, B, C
#>  B, C
#>  C, C, C
#>  C, C
## data.frame format
contexts(dts_tree, sequence = TRUE)
#>    context
#> 1     A, A
#> 2     B, A
#> 3  B, C, A
#> 4     C, A
#> 5     A, B
#> 6  B, B, B
#> 7  C, B, B
#> 8     B, B
#> 9  B, C, B
#> 10 C, C, B
#> 11    C, B
#> 12 C, A, C
#> 13    A, C
#> 14 A, B, C
#> 15 C, B, C
#> 16    B, C
#> 17 C, C, C
#> 18    C, C
contexts(dts_tree, frequency = "total")
#>    context freq
#> 1     A, A    8
#> 2     B, A    8
#> 3  B, C, A    7
#> 4     C, A   11
#> 5     A, B   11
#> 6  B, B, B    5
#> 7  C, B, B    5
#> 8     B, B   13
#> 9  B, C, B    5
#> 10 C, C, B    5
#> 11    C, B   13
#> 12 C, A, C    5
#> 13    A, C    8
#> 14 A, B, C    6
#> 15 C, B, C    6
#> 16    B, C   15
#> 17 C, C, C    5
#> 18    C, C   11
contexts(dts_tree, frequency = "detailed")
#>    context freq A B C
#> 1     A, A    8 4 4 0
#> 2     B, A    8 1 4 3
#> 3  B, C, A    7 0 3 4
#> 4     C, A   11 3 3 5
#> 5     A, B   11 2 3 6
#> 6  B, B, B    5 2 2 1
#> 7  C, B, B    5 2 2 1
#> 8     B, B   13 4 5 4
#> 9  B, C, B    5 1 1 3
#> 10 C, C, B    5 1 2 2
#> 11    C, B   13 2 5 6
#> 12 C, A, C    5 2 1 2
#> 13    A, C    8 2 3 3
#> 14 A, B, C    6 4 0 2
#> 15 C, B, C    6 2 3 1
#> 16    B, C   15 7 5 3
#> 17 C, C, C    5 1 2 2
#> 18    C, C   11 1 5 5