Text based representation of a vlmc

This function 'draws' a context tree as a text.

Usage

# S3 method for class 'vlmc'
draw(ct, control = draw_control(), prob = TRUE, ...)

# S3 method for class 'vlmc_cpp'
draw(ct, control = draw_control(), prob = TRUE, ...)

Arguments

ct

a fitted vlmc.

control

a list of low level control parameters of the text representation. See details and draw_control().

prob

this parameter controls the display of node level information in the tree. The default prob=TRUE represents the conditional distribution of the states given the (partial) context associated to the node. Setting prob=FALSE replaces the conditional distribution by the frequency of the states that follow the context as in draw.ctx_tree(). Setting prob=NULL removes all additional information.

...

additional arguments for draw.

Value

the context tree (invisibly).

Details

The function uses basic "ascii art" to represent the context tree. Characters used to represent the structure of the tree, e.g. branches, can be modified using draw_control().

In addition to the structure of the context tree, draw can represent information attached to the node (contexts and partial contexts). This is controlled by additional parameters depending on the type of the context tree.

Examples

dts <- sample(c("A", "B", "C"), 500, replace = TRUE)
model <- vlmc(dts, alpha = 0.05)
draw(model)
#> * (0.344, 0.3, 0.356)
draw(model, prob = FALSE)
#> * (172, 150, 178)
draw(model, prob = NULL)
#> *