Predictive quality metrics for VLMC

This function computes and returns predictive quality metrics for VLMC and VLMC with covariates.

Usage

# S3 method for vlmc
metrics(model, ...)

# S3 method for metrics.vlmc
print(x, ...)

Arguments

model

The context based model on which to compute predictive metrics.

...

Additional parameters for predictive metrics computation.

x

A metrics.vlmc object, results of a call to metrics.vlmc()

Value

An object of class metrics.vlmc with the following components:

accuracy: the accuracy of the predictions

conf_mat: the confusion matrix of the predictions, with predicted values in rows and true values in columns

auc: the AUC of the predictive model

The object has a print method that recalls basic information about the model together with the values of the components above.

Details

A context based model computes transition probabilities for its contexts. Using a maximum transition probability decision rule, this can be used to "predict" the new state that is the more likely to follow the current one, given the context. The quality of these predictions is evaluated using standard metrics including:

accuracy

the full confusion matrix

the area under the roc curve (AUC), considering the context based model as a (conditional) probability estimator. We use Hand and Till (2001) multiclass AUC in case of a state space with more than 2 states

Methods (by generic)

print(metrics.vlmc): Prints the predictive metrics of the VLMC model.

References

David J. Hand and Robert J. Till (2001). A Simple Generalisation of the Area Under the ROC Curve for Multiple Class Classification Problems. Machine Learning 45(2), p. 171--186. DOI: 10.1023/A:1010920819831.

See also