mixvlmc: Variable Length Markov Chains with Covariates

Estimates Variable Length Markov Chains (VLMC) models and VLMC with covariates models from discrete sequences. Supports model selection via information criteria and simulation of new sequences from an estimated model. See Bühlmann, P. and Wyner, A. J. (1999) doi:10.1214/aos/1018031204 for VLMC and Zanin Zambom, A., Kim, S. and Lopes Garcia, N. (2022) doi:10.1111/jtsa.12615 for VLMC with covariates.

Package options

Mixvlmc uses the following options():

mixvlmc.maxit: maximum number of iterations in model fitting for covlmc()

mixvlmc.predictive: specifies the computing engine used for model fitting for covlmc(). Two values are supported:

"glm" (default value): covlmc() uses stats::glm() with a binomial link (stats::binomial()) for a two values state space, and VGAM::vglm() with a multinomial link (VGAM::multinomial()) for a state space with three or more values;

"multinom": covlmc() uses nnet::multinom() in all cases.

The first option "glm" is recommended as both stats::glm() and VGAM::vglm() are able to detect and deal with degeneracy in the data set.

mixvlmc.backend: specifies the implementation used for the context tree construction in ctx_tree(), vlmc() and tune_vlmc(). Two values are supported:

"R" (default value): this corresponds to the original almost pure R implementation.

"C++": this corresponds to the experimental C++ implementation. This version is significantly faster than the R version, but is still considered experimental.

See also

Author

Maintainer: Fabrice Rossi Fabrice.Rossi@apiacoa.org (ORCID) [copyright holder]

Other contributors:

Hugo Le Picard lepicardhugo@gmail.com (ORCID) [contributor]

Guénolé Joubioux guenole.joubioux@gmail.com [contributor]