Simulate a discrete time series for a covlmc

This function simulates a time series from the distribution estimated by the given covlmc object.

Usage

# S3 method for covlmc
simulate(object, nsim = 1, seed = NULL, covariate, init = NULL, ...)

Arguments

object

a fitted covlmc object.

nsim

length of the simulated time series (defaults to 1).

seed

an optional random seed.

covariate

values of the covariates

init

an optional initial sequence for the time series

...

additional arguments.

Details

A VLMC with covariates model needs covariates to compute its transition probabilities. The covariates must be submitted as a data frame using the covariate argument. In addition, the time series can be initiated by a fixed sequence specified via the init parameter.

Examples

pc <- powerconsumption[powerconsumption$week == 5, ]
dts <- cut(pc$active_power, breaks = c(0, quantile(pc$active_power, probs = c(0.5, 1))))
dts_cov <- data.frame(day_night = (pc$hour >= 7 & pc$hour <= 17))
m_cov <- covlmc(dts, dts_cov, min_size = 5)
# new week with day light from 6:00 to 18:00
new_cov <- data.frame(day_night = rep(c(rep(FALSE, 59), rep(TRUE, 121), rep(FALSE, 60)), times = 7))
new_dts <- simulate(m_cov, nrow(new_cov), seed = 0, covariate = new_cov)
new_dts_2 <- simulate(m_cov, nrow(new_cov), seed = 0, covariate = new_cov, init = dts[1:10])