Contexts of a VLMC with covariates

This function returns the different contexts present in a VLMC with covariates, possibly with some associated data.

Usage

# S3 method for covlmc
contexts(
  ct,
  type = c("auto", "list", "data.frame"),
  reverse = TRUE,
  frequency = NULL,
  counts = c("desc", "local"),
  model = NULL,
  hsize = FALSE,
  metrics = FALSE,
  merging = FALSE,
  ...
)

Arguments

ct

a fitted covlmc model.

type

result type (see details).

reverse

logical (defaults to FALSE). See details.

frequency

specifies the counts to be included in the result data.frame. The default value of NULL does not include anything. "total" gives the number of occurrences of each context in the original sequence. "detailed" includes in addition the break down of these occurrences into all the possible states.

counts

specifies how the counts reported by frequency are computed. The default value "desc" includes both counts that are specific to the context (if any) and counts from the descendants of the context in the tree. When counts = "local" the counts include only the number of times the context appears without being the last part of a longer context.

model

specifies whether to include the model associated to a each context. The default result with model=NULL does not include any model. Setting model to "coef" adds the coefficients of the models in a coef column, while "full" include the models themselves (as R objects) in a model column.

hsize

if TRUE, adds a hsize column to the result data frame that gives for each context the size of the history of covariates used by the model.

metrics

if TRUE, adds predictive metrics for each context (see metrics() for the definition of predictive metrics).

merging

if TRUE, adds a merged column to the result data frame. For a normal context, the value of merged is FALSE. Contexts that share the same model have a TRUE merged value.

...

additional arguments for the contexts function.

Value

the list of the contexts represented in this tree or a data.frame with more content.

Details

The result is a list of all contexts when type="auto" (or type="list"), and no details have been asked via specific parameters (e.g. by setting model to a non NULL value).

Other results are obtained only with type="data.frame" (or type="auto"). See contexts.ctx_tree() for details about the frequency parameter. When model is non NULL, the resulting data.frame contains the models associated to each context (either the full R model or its coefficients).

State order in a context

Notice that contexts are given by default in the "reverse" order used by the VLMC papers: older values are on the right. For instance, the context c(0, 1) is reported if the sequence 1, then 0 appeared in the time series used to build the context tree. Set reverse to FALSE for the reverse convention.

See also

Examples

pc <- powerconsumption[powerconsumption$week == 5, ]
breaks <- c(0, median(pc$active_power), max(pc$active_power))
dts <- cut(pc$active_power, breaks = breaks)
dts_cov <- data.frame(day_night = (pc$hour >= 7 & pc$hour <= 17))
m_cov <- covlmc(dts, dts_cov, min_size = 5)
contexts(m_cov, model = "coef")
#>        context         coef
#> 1 (0,1.34].... -2.81304....
#> 2 (0,1.34].... -19.5660....
#> 3 (0,1.34].... -1.85629....
#> 4 (1.34,7..... 1.642227....
#> 5 (1.34,7..... 2.643655....
contexts(m_cov, model = "full")
#>        context        model
#> 1 (0,1.34].... c(`(Inte....
#> 2 (0,1.34].... c(`(Inte....
#> 3 (0,1.34].... c(`(Inte....
#> 4 (1.34,7..... c(`(Inte....
#> 5 (1.34,7..... c(`(Inte....