Contexts of a VLMC

This function extracts all the contexts from a fitted VLMC, possibly with some associated data.

Usage

# S3 method for vlmc
contexts(
  ct,
  type = c("auto", "list", "data.frame"),
  reverse = TRUE,
  frequency = NULL,
  positions = FALSE,
  counts = c("desc", "local"),
  cutoff = NULL,
  metrics = FALSE,
  ...
)

Arguments

ct

a context tree.

type

result type (see details).

reverse

logical (defaults to TRUE). See details.

frequency

specifies the counts to be included in the result data.frame. The default value of NULL does not include anything. "total" gives the number of occurrences of each context in the original sequence. "detailed" includes in addition the break down of these occurrences into all the possible states.

positions

logical (defaults to FALSE). Specify whether the positions of each context in the time series used to build the context tree should be reported in a positions column of the result data frame. The availability of the positions depends on the way the context tree was built. See details for the definition of a position.

counts

specifies how the counts reported by frequency are computed. The default value "desc" includes both counts that are specific to the context (if any) and counts from the descendants of the context in the tree. When counts = "local" the counts include only the number of times the context appears without being the last part of a longer context.

cutoff

specifies whether to include the cut off value associated to each context (see cutoff() and prune()). The default result with cutoff=NULL does not include those values. Setting cutoff to quantile adds the cut off values in quantile scale, while cutoff="native" adds them in the native scale. The returned values are directly based on the log likelihood ratio computed in the context tree and are not modified to ensure pruning (as when cutoff() is called by raw=TRUE).

metrics

if TRUE, adds predictive metrics for each context (see metrics() for the definition of predictive metrics).

...

additional arguments for the contexts function.

Value

The list of the contexts represented in this tree or a data.frame with more content.

Details

The default result for type="auto" (or type="list"), frequency=NULL, cutoff=NULL and metrics=FALSE is the list of all contexts.

Other results are obtained only with type="auto" or type="data.frame". See contexts.ctx_tree() for details about the frequency parameter. When cutoff is non NULL, the resulting data.frame contains a cutoff column with the cut off values, either in quantile or in native scale. See cutoff() and prune() for the definitions of cut off values and of the two scales.

Cut off values

The cut off values reported by contexts.vlmc can be different from the ones reported by cutoff() for three reasons:

cutoff() reports only useful cut off values, i.e., cut off values that should induce a simplification of the VLMC when used in prune(). This exclude cut off values associated to simple contexts that are smaller than the ones of their descendants in the context tree. Those values are reported by context.vlmc.

context.vlmc reports only cut off values of actual contexts, while cutoff() reports cut off values for all nodes of the context tree.

values are not modified to induce pruning, contrarily to the default behaviour of cutoff()

Positions

A position of a context ctx in the time series x is an index value t such that the context ends with x[t]. Thus x[t+1] is after the context. For instance if x=c(0, 0, 1, 1) and ctx=c(0, 1) (in standard state order), then the position of ctx in x is 3.

State order in a context

Notice that contexts are given by default in the "reverse" order used by the VLMC papers: older values are on the right. For instance, the context c(0, 1) is reported if the sequence 1, then 0 appeared in the time series used to build the context tree. Set reverse to FALSE for the reverse convention.

See also

Examples

dts <- sample(as.factor(c("A", "B", "C")), 100, replace = TRUE)
model <- vlmc(dts, alpha = 0.5)
contexts(model)
#> [[1]]
#> [1] A A A C A
#> Levels: A B C
#> 
#> [[2]]
#> [1] A A A C B
#> Levels: A B C
#> 
#> [[3]]
#> [1] A A A C
#> Levels: A B C
#> 
#> [[4]]
#> [1] A A A
#> Levels: A B C
#> 
#> [[5]]
#> [1] A A B A
#> Levels: A B C
#> 
#> [[6]]
#> [1] A A B
#> Levels: A B C
#> 
#> [[7]]
#> [1] A A C
#> Levels: A B C
#> 
#> [[8]]
#> [1] A B A
#> Levels: A B C
#> 
#> [[9]]
#> [1] A B B A
#> Levels: A B C
#> 
#> [[10]]
#> [1] A B B
#> Levels: A B C
#> 
#> [[11]]
#> [1] A B C
#> Levels: A B C
#> 
#> [[12]]
#> [1] A C A B
#> Levels: A B C
#> 
#> [[13]]
#> [1] A C A
#> Levels: A B C
#> 
#> [[14]]
#> [1] A C B A B
#> Levels: A B C
#> 
#> [[15]]
#> [1] A C B A
#> Levels: A B C
#> 
#> [[16]]
#> [1] A C B
#> Levels: A B C
#> 
#> [[17]]
#> [1] A C C
#> Levels: A B C
#> 
#> [[18]]
#> [1] B A B B
#> Levels: A B C
#> 
#> [[19]]
#> [1] B A B C
#> Levels: A B C
#> 
#> [[20]]
#> [1] B A B
#> Levels: A B C
#> 
#> [[21]]
#> [1] B A C C
#> Levels: A B C
#> 
#> [[22]]
#> [1] B A C
#> Levels: A B C
#> 
#> [[23]]
#> [1] B A
#> Levels: A B C
#> 
#> [[24]]
#> [1] B B A
#> Levels: A B C
#> 
#> [[25]]
#> [1] B B B
#> Levels: A B C
#> 
#> [[26]]
#> [1] B B
#> Levels: A B C
#> 
#> [[27]]
#> [1] B C A C
#> Levels: A B C
#> 
#> [[28]]
#> [1] B C A
#> Levels: A B C
#> 
#> [[29]]
#> [1] B C B
#> Levels: A B C
#> 
#> [[30]]
#> [1] B C C
#> Levels: A B C
#> 
#> [[31]]
#> [1] C A C A
#> Levels: A B C
#> 
#> [[32]]
#> [1] C A C
#> Levels: A B C
#> 
#> [[33]]
#> [1] C A
#> Levels: A B C
#> 
#> [[34]]
#> [1] C B A A
#> Levels: A B C
#> 
#> [[35]]
#> [1] C B A
#> Levels: A B C
#> 
#> [[36]]
#> [1] C B B
#> Levels: A B C
#> 
#> [[37]]
#> [1] C B C
#> Levels: A B C
#> 
#> [[38]]
#> [1] C C B A
#> Levels: A B C
#> 
#> [[39]]
#> [1] C C B C A
#> Levels: A B C
#> 
#> [[40]]
#> [1] C C B C
#> Levels: A B C
#> 
#> [[41]]
#> [1] C C B
#> Levels: A B C
#> 
#> [[42]]
#> [1] C C C B C
#> Levels: A B C
#> 
#> [[43]]
#> [1] C C C B
#> Levels: A B C
#> 
#> [[44]]
#> [1] C C C
#> Levels: A B C
#> 
#> [[45]]
#> [1] C C
#> Levels: A B C
#> 
contexts(model, frequency = "total")
#>         context freq
#> 1  A, A, A,....    2
#> 2  A, A, A,....    2
#> 3    A, A, A, C    4
#> 4       A, A, A    6
#> 5    A, A, B, A    2
#> 6       A, A, B    3
#> 7       A, A, C    4
#> 8       A, B, A    3
#> 9    A, B, B, A    2
#> 10      A, B, B    4
#> 11      A, B, C    4
#> 12   A, C, A, B    2
#> 13      A, C, A    5
#> 14 A, C, B,....    2
#> 15   A, C, B, A    3
#> 16      A, C, B    4
#> 17      A, C, C    2
#> 18   B, A, B, B    2
#> 19   B, A, B, C    2
#> 20      B, A, B    5
#> 21   B, A, C, C    2
#> 22      B, A, C    4
#> 23         B, A   13
#> 24      B, B, A    4
#> 25      B, B, B    3
#> 26         B, B    9
#> 27   B, C, A, C    2
#> 28      B, C, A    4
#> 29      B, C, B    2
#> 30      B, C, C    3
#> 31   C, A, C, A    2
#> 32      C, A, C    4
#> 33         C, A   10
#> 34   C, B, A, A    2
#> 35      C, B, A    6
#> 36      C, B, B    2
#> 37      C, B, C    3
#> 38   C, C, B, A    2
#> 39 C, C, B,....    2
#> 40   C, C, B, C    3
#> 41      C, C, B    5
#> 42 C, C, C,....    2
#> 43   C, C, C, B    3
#> 44      C, C, C    4
#> 45         C, C   10
contexts(model, cutoff = "quantile")
#>         context     cutoff
#> 1  A, A, A,.... 0.25000000
#> 2  A, A, A,.... 0.25000000
#> 3    A, A, A, C 0.59259259
#> 4       A, A, A 0.30931243
#> 5    A, A, B, A 0.44444444
#> 6       A, A, B 0.88484297
#> 7       A, A, C 0.04537656
#> 8       A, B, A 0.22821187
#> 9    A, B, B, A 0.50000000
#> 10      A, B, B 0.98353938
#> 11      A, B, C 0.24588484
#> 12   A, C, A, B 0.32000000
#> 13      A, C, A 0.80375514
#> 14 A, C, B,.... 0.44444444
#> 15   A, C, B, A 0.84375000
#> 16      A, C, B 0.79012346
#> 17      A, C, C 0.11111111
#> 18   B, A, B, B 0.48000000
#> 19   B, A, B, C 0.48000000
#> 20      B, A, B 0.80798722
#> 21   B, A, C, C 0.25000000
#> 22      B, A, C 0.64535555
#> 23         B, A 0.59505147
#> 24      B, B, A 0.35116598
#> 25      B, B, B 0.14814815
#> 26         B, B 0.68445872
#> 27   B, C, A, C 0.50000000
#> 28      B, C, A 0.70233196
#> 29      B, C, B 0.39506173
#> 30      B, C, C 0.44444444
#> 31   C, A, C, A 0.25000000
#> 32      C, A, C 0.51200000
#> 33         C, A 0.27045625
#> 34   C, B, A, A 0.22222222
#> 35      C, B, A 0.78032876
#> 36      C, B, B 0.26446281
#> 37      C, B, C 0.09391435
#> 38   C, C, B, A 0.48000000
#> 39 C, C, B,.... 0.44444444
#> 40   C, C, B, C 0.48600000
#> 41      C, C, B 0.66666667
#> 42 C, C, C,.... 0.44444444
#> 43   C, C, C, B 0.84375000
#> 44      C, C, C 0.69120000
#> 45         C, C 0.76223750