Text based representation of a covlmc model

This function 'draws' a covlmc as a text.

Usage

# S3 method for covlmc
draw(
  ct,
  format,
  control = draw_control(),
  model = c("coef", "full"),
  p_value = FALSE,
  with_state = FALSE,
  constant_as_prob = TRUE,
  ...
)

Arguments

ct

a fitted covlmc model.

format

a character string that specifies the output format of the function. Possible values are ascii (default) and latex. See details.

control

a list of low level control parameters of the text representation. See details and draw_control().

model

this parameter controls the display of logistic models associated to nodes (accepted values: "coeff", "full" and NULL). The interpretation of the parameter depends on the format, see below for details.

p_value

specifies whether the p-values of the likelihood ratio tests conducted during the covlmc construction must be included in the representation (defaults to FALSE).

with_state

specifies whether to display the state associated to each dimension of the logistic model (see details).

constant_as_prob

specifies how to represent constant logistic models for format="ascii" (defaults to TRUE, see details). Disregarded when format="latex".

...

additional arguments for draw.

Details

The function uses different text based formats (plain "ascii art" and LaTeX) to represent the context tree. Fine tuning of the representation can be done via the draw_control() function.

Contrarily to draw() functions adapted to context trees draw.ctx_tree() and VLMC draw.vlmc(), the present function does not try to produce similar results for the "ascii" format and the "latex" format as the "ascii" format is intrinsically more limited in terms of model representations. This is detailed below.

Format

The format parameter specifies the format used for the textual output. With the default value ascii the output is produced in "ascii art" using by default only ascii characters (notice that draw_control() can be used to specified non ascii characters, but this is discouraged).

With the latex value, the output is produced in LaTeX, leveraging the forest Latex package (see https://ctan.org/pkg/forest). Each call to draw.covlmc() produces a full forest LaTeX environment. This can be included as is in a LaTeX document, provided the forest package is loaded in the preamble of the document. The LaTeX output is sanitized to avoid potential problems induced by special characters in the names of the states of the context tree.

"ascii" format

Parameters

When format="ascii" the parameters are interpreted as follows:

model: the default model="coef" represents only the coefficients of the logistic models associated to each context. model="full" includes the name of the variables in the representation. Setting model=NULL removes the model representations. Additional parameters can be used to tweak model representations (see below).

constant_as_prob: specifies whether to represent logistic models that do not use covariates (a.k.a. constant models) using the probability distributions they induce on the state space (default behaviour with constant_as_prob=TRUE) or as normal models (when set to FALSE). This is not taken into account when model is not set to "coef".

fields of the control list:

intercept_sep: character(s) used to separate the intercept from the other coefficients in model representation.

time_sep: character(s) used to split the coefficients list by blocks associated to time delays in the covariate inclusion into the logistic model. The first block contains the intercept(s), the second block the covariate values a time t-1, the third block at time t-2, etc.

level_sep: character(s) used separate levels from model, see below.

open_p_value and close_p_value: delimiters used around the p-values when p_value=TRUE

State representation

When model is not NULL, the coefficients of the logistic models are presented, organized in rows associated to states. One state is used as the reference state and the logistic model aims at predicting the ratio of probability between another state and the reference one (in log scale). When with_state is TRUE, the display includes for each row of coefficients the target state. This is useful when using e.g. VGAM::vglm() as unused levels of the target variable will be automatically dropped from the model, leading to a reduce number of rows. The reference state is either shown on the first row if model is "full" or after the state on each row if model is "coef". States are separated from the model representation by the character(s) specified in level_sep in the control list.

"latex" format

Parameters

When format="latex" the parameters are interpreted as follows:

model: the models are always represented completely in the LaTeX export unless model is set to NULL.

constant_as_prob: in the LaTeX export, constant logistic models are always represented by the corresponding probability distribution on the state space, regardless of the value of constant_as_prob.

fields of the control list:

orientation: specifies the orientation of the tree, either the default "vertical" (expanding from top to bottom) or "horizontal" (expanding from right to left);

tab_orientation: specifies the orientation of the tables used to represent model coefficients in the tree, either the default "vertical" (covariates are listed on one column) or "horizontal" (covariates are listed on one row);

fontsize and prob_fontsize handle the size of the fonts used for the states and for the models, see draw_control() for details;

decoration can be used to add borders around states, see draw_control() for details;

State representation

When model is not NULL, the coefficients of the logistic models are presented, organized in rows or in columns (depending tab_orientation) on associated to states. One state is used as the reference state and the logistic model aims at predicting the ratio of probability between another state and the reference one (in log scale). When with_state is TRUE, the display includes for each row/column of coefficients the target state. The reference state is shown on the first row/column.

Variable representation

When the representation includes the names of the variables used by the logistic models, they are the one generated by the underlying logistic model, e.g. stats::glm(). Numerical variable names are used as is, while factors have levels appended. The intercept is denoted (I) to save space.

When format="ascii", the time delays are represented by an underscore followed by the time delay. For instance if the model uses the numerical covariate y with two delays, it will appear with two variables y_1 and y_2.

When format="latex", the representation uses a temporal subscript of the form t-1, t-2, etc.

Examples

pc <- powerconsumption[powerconsumption$week == 5, ]
dts <- cut(pc$active_power,
  breaks = c(0, quantile(pc$active_power, probs = c(0.5, 1)))
)
dts_cov <- data.frame(day_night = (pc$hour >= 7 & pc$hour <= 17))
m_cov <- covlmc(dts, dts_cov, min_size = 5)
draw(m_cov, control = draw_control(digits = 3))
#> *
#> +-- (0,1.34]
#> |   +-- (0,1.34]
#> |   |   +-- (0,1.34] ([ -2.81 & 2.18 | -4.45 | 2.42 ])
#> |   |   '-- (1.34,7.54] ([ -19.6 & 18.4 ])
#> |   '-- (1.34,7.54] (0.865, 0.135)
#> '-- (1.34,7.54] (0.0734, 0.927)
draw(m_cov, model = NULL)
#> *
#> +-- (0,1.34]
#> |   +-- (0,1.34]
#> |   |   +-- (0,1.34]
#> |   |   '-- (1.34,7.54]
#> |   '-- (1.34,7.54]
#> '-- (1.34,7.54]
draw(m_cov, p_value = TRUE)
#> *
#> +-- (0,1.34]
#> |   +-- (0,1.34]
#> |   |   +-- (0,1.34] (<0.02329> [ -2.813 & 2.178 | -4.449 | 2.418 ])
#> |   |   '-- (1.34,7.54] (<0.01834> [ -19.57 & 18.39 ])
#> |   '-- (1.34,7.54] (<0.6763> 0.8649, 0.1351)
#> '-- (1.34,7.54] (<0.8175> 0.07341, 0.9266)
draw(m_cov, p_value = FALSE, control = draw_control(time_sep = " ^ "))
#> *
#> +-- (0,1.34]
#> |   +-- (0,1.34]
#> |   |   +-- (0,1.34] ([ -2.813 & 2.178 ^ -4.449 ^ 2.418 ])
#> |   |   '-- (1.34,7.54] ([ -19.57 & 18.39 ])
#> |   '-- (1.34,7.54] (0.8649, 0.1351)
#> '-- (1.34,7.54] (0.07341, 0.9266)
draw(m_cov, model = "full", control = draw_control(time_sep = " ^ "))
#> *
#> +-- (0,1.34]
#> |   +-- (0,1.34]
#> |   |   +-- (0,1.34] ([ (I)    & day_night_1TRUE ^ day_night_2TRUE ^ day_night_3TRUE
#> |   |   |               -2.813 & 2.178           ^ -4.449          ^ 2.418           ])
#> |   |   '-- (1.34,7.54] ([ (I)    & day_night_1TRUE
#> |   |                      -19.57 & 18.39           ])
#> |   '-- (1.34,7.54] ([ (I)   
#> |                      -1.856 ])
#> '-- (1.34,7.54] ([ (I)  
#>                    2.535 ])
draw(m_cov, format = "latex", control = draw_control(orientation = "h"))
#> \begin{forest}
#>   for tree={grow'=west}[\textbf{$\epsilon$}
#> [\textbf{(0,1.34]}
#> [\textbf{(0,1.34]}
#> [\multicolumn{2}{c}{\textbf{(0,1.34]}}\\\hline
#> {\small (I)}&{\small -2.813}\\
#> {\small $\text{day\_night}_{t-1}$}&{\small 2.178}\\
#> {\small $\text{day\_night}_{t-2}$}&{\small -4.449}\\
#> {\small $\text{day\_night}_{t-3}$}&{\small 2.418}\\, align=cc]
#> [\multicolumn{2}{c}{\textbf{(1.34,7.54]}}\\\hline
#> {\small (I)}&{\small -19.57}\\
#> {\small $\text{day\_night}_{t-1}$}&{\small 18.39}\\, align=cc]
#> ]
#> [\multicolumn{2}{c}{\textbf{(1.34,7.54]}}\\\hline
#> {\small 0.8649}&{\small 0.1351}, align=cc]
#> ]
#> [\multicolumn{2}{c}{\textbf{(1.34,7.54]}}\\\hline
#> {\small 0.07341}&{\small 0.9266}, align=cc]
#> ]
#> \end{forest}