Text based representation of a context tree

This function 'draws' a context tree as a text.

Usage

# S3 method for ctx_tree_cpp
draw(ct, format, control = draw_control(), frequency = NULL, ...)

# S3 method for ctx_tree
draw(ct, format, control = draw_control(), frequency = NULL, ...)

Arguments

ct

a context tree.

format

a character string that specifies the output format of the function. Possible values are "text" (default) and "latex". See details.

control

a list of low level control parameters of the text representation. See details and draw_control().

frequency

this parameter controls the display of node level information in the tree. The default NULL value does not include anything. Setting frequency to "total" includes the frequency of the (partial) context of the node, while "detailed" includes the frequency of the states that follow the context (as in contexts.ctx_tree()).

...

additional arguments for draw.

Value

the context tree (invisibly).

Details

The function uses different text based formats (plain "ascii art" and LaTeX) to represent the context tree. Fine tuning of the representation can be done via the draw_control() function.

In addition to the structure of the context tree, draw() can represent information attached to the nodes (contexts and partial contexts). This is controlled by additional parameters depending on the type of the context tree. In general, parameters given directly to draw() specify what information is represented while details on how this representation is made can be controlled via the control parameter and the associated draw_control() function.

Format

The format parameter specifies the format used for the textual output. With the default value "text" the output is produced in "ascii art" using by default only ascii characters (notice that draw_control() can be used to specified non ascii characters, but this is discouraged).

With the latex value, the output is produced in LaTeX, leveraging the forest Latex package (see https://ctan.org/pkg/forest). Each call to draw() produces a full forest LaTeX environment. This can be included as is in a LaTeX document, provided the forest package is loaded in the preamble of the document. The LaTeX output is sanitized to avoid potential problems induced by special characters in the names of the states of the context tree.

Examples

rdts_c <- sample(c("A", "B", "CD"), 100, replace = TRUE)
ctree_c <- ctx_tree(rdts_c, min_size = 10, max_depth = 2)
draw(ctree_c, frequency = "total")
#> ▪ (100)
#> ├─ A (38)
#> │  ├─ A (17)
#> │  └─ B (14)
#> ├─ B (29)
#> │  └─ CD (13)
#> └─ CD (32)
#>    ├─ A (12)
#>    └─ CD (12)
draw(ctree_c, frequency = "detailed")
#> ▪ (39,29,32)
#> ├─ A (17,9,12)
#> │  ├─ A (7,4,6)
#> │  └─ B (6,4,4)
#> ├─ B (14,7,8)
#> │  └─ CD (6,4,3)
#> └─ CD (7,13,12)
#>    ├─ A (2,6,4)
#>    └─ CD (4,4,4)
## LaTeX output
draw(ctree_c, "latex", frequency = "detailed")
#> \begin{forest}
#> [\multicolumn{3}{c}{\textbf{$\epsilon$}}\\\hline
#> {\small 39}&{\small 29}&{\small 32}, align=ccc
#> [\multicolumn{3}{c}{\textbf{A}}\\\hline
#> {\small 17}&{\small 9}&{\small 12}, align=ccc
#> [\multicolumn{3}{c}{\textbf{A}}\\\hline
#> {\small 7}&{\small 4}&{\small 6}, align=ccc]
#> [\multicolumn{3}{c}{\textbf{B}}\\\hline
#> {\small 6}&{\small 4}&{\small 4}, align=ccc]
#> ]
#> [\multicolumn{3}{c}{\textbf{B}}\\\hline
#> {\small 14}&{\small 7}&{\small 8}, align=ccc
#> [\multicolumn{3}{c}{\textbf{CD}}\\\hline
#> {\small 6}&{\small 4}&{\small 3}, align=ccc]
#> ]
#> [\multicolumn{3}{c}{\textbf{CD}}\\\hline
#> {\small 7}&{\small 13}&{\small 12}, align=ccc
#> [\multicolumn{3}{c}{\textbf{A}}\\\hline
#> {\small 2}&{\small 6}&{\small 4}, align=ccc]
#> [\multicolumn{3}{c}{\textbf{CD}}\\\hline
#> {\small 4}&{\small 4}&{\small 4}, align=ccc]
#> ]
#> ]
#> \end{forest}
rdts_c <- sample(c("A$", "_{B", "{C}_{D}"), 100, replace = TRUE)
ctree_c <- ctx_tree(rdts_c, min_size = 10, max_depth = 2)
## the LaTeX output is sanitized
draw(ctree_c, "latex", frequency = "detailed")
#> \begin{forest}
#> [\multicolumn{3}{c}{\textbf{$\epsilon$}}\\\hline
#> {\small 32}&{\small 30}&{\small 38}, align=ccc
#> [\multicolumn{3}{c}{\textbf{A\$}}\\\hline
#> {\small 10}&{\small 6}&{\small 16}, align=ccc
#> [\multicolumn{3}{c}{\textbf{A\$}}\\\hline
#> {\small 3}&{\small 1}&{\small 6}, align=ccc]
#> [\multicolumn{3}{c}{\textbf{\_\{B}}\\\hline
#> {\small 1}&{\small 2}&{\small 8}, align=ccc]
#> [\multicolumn{3}{c}{\textbf{\{C\}\_\{D\}}}\\\hline
#> {\small 6}&{\small 2}&{\small 2}, align=ccc]
#> ]
#> [\multicolumn{3}{c}{\textbf{\_\{B}}\\\hline
#> {\small 11}&{\small 10}&{\small 8}, align=ccc
#> [\multicolumn{3}{c}{\textbf{\_\{B}}\\\hline
#> {\small 3}&{\small 5}&{\small 2}, align=ccc]
#> [\multicolumn{3}{c}{\textbf{\{C\}\_\{D\}}}\\\hline
#> {\small 6}&{\small 5}&{\small 2}, align=ccc]
#> ]
#> [\multicolumn{3}{c}{\textbf{\{C\}\_\{D\}}}\\\hline
#> {\small 10}&{\small 14}&{\small 14}, align=ccc
#> [\multicolumn{3}{c}{\textbf{A\$}}\\\hline
#> {\small 4}&{\small 6}&{\small 6}, align=ccc]
#> [\multicolumn{3}{c}{\textbf{\{C\}\_\{D\}}}\\\hline
#> {\small 3}&{\small 6}&{\small 5}, align=ccc]
#> ]
#> ]
#> \end{forest}