Control parameters for draw

This function returns a list used to fine tune the draw() function behaviour.

Usage

draw_control(
  root = "*",
  first_node = "+",
  next_node = "'",
  vbranch = "|",
  hbranch = "--",
  open_ct = "(",
  close_ct = ")"
)

Arguments

root

character used for the root node.

first_node

characters used for the first child of a node.

next_node

characters used for other children of a node.

vbranch

characters used to represent a branch in a vertical way.

hbranch

characters used to represent a branch in a horizontal was.

open_ct

characters used to start each node specific text representation.

close_ct

characters used to end each node specific text representation.

Value

a list

Examples

draw_control(open_ct = "[", close_ct = "]")
#> $root
#> [1] "*"
#> 
#> $first_node
#> [1] "+"
#> 
#> $next_node
#> [1] "'"
#> 
#> $vbranch
#> [1] "|"
#> 
#> $hbranch
#> [1] "--"
#> 
#> $open_ct
#> [1] "["
#> 
#> $close_ct
#> [1] "]"
#>